requirements.txt
autopep8==2.0.1
black==23.1.0
cffi==1.15.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.7.0
mypy-extensions==1.0.0
networkx==3.0
numpy==1.24.2
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
pycodestyle==2.10.0
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pywfc==0.0.0
PyYAML==6.0
scipy==1.10.0
six==1.16.0
tcod==15.0.0
typing_extensions==4.4.0
